move definitonFile.ts
